+++ backend/myserver/package.json
@@ -32,7 +32,6 @@
     "@nestjs/platform-express": "^8.0.0",
     "@nestjs/platform-socket.io": "^8.2.3",
     "@nestjs/serve-static": "^2.2.2",
-    "@nestjs/swagger": "^5.1.5",
     "@nestjs/typeorm": "^8.0.2",
     "@nestjs/websockets": "^8.2.3",
     "@socket.io/admin-ui": "^0.2.0",
@@ -52,7 +51,7 @@
     "otplib": "^12.0.1",
     "passport": "^0.4.1",
     "passport-42": "^1.2.6",
-    "passport-custom": "^1.1.1",
+    "passport-google-oauth2": "^0.2.0",
     "passport-jwt": "^4.0.0",
     "pg": "^8.7.1",
     "qrcode": "^1.5.0",
@@ -60,7 +59,6 @@
     "rimraf": "^3.0.2",
     "rxjs": "^7.2.0",
     "socket.io": "^4.2.0",
-    "swagger-ui-express": "^4.1.6",
     "tslib": "^2.3.1",
     "typeorm": "^0.2.37"
   },
