+++ backend/src/.env
@@ -1,4 +1,4 @@
-DB_HOST=localhost
+DB_HOST=10.19.218.78
 DB_PORT=5432
 DB_USER_NAME=postgres
 DB_PASSWORD=12345
