--- backend/src/auth/signup.validationpipe.ts
@@ -1,22 +1,0 @@
-import { ArgumentMetadata, BadRequestException, Injectable, PipeTransform } from "@nestjs/common";
-import { plainToClass } from "class-transformer";
-import { validate } from "class-validator";
-
-@Injectable()
-export class validationPipe implements PipeTransform {
-    async transform(value : any, {metatype} : ArgumentMetadata) {
-        if (!metatype || !this.toValidate(metatype))
-            return value;
-
-        const object = plainToClass(metatype, value);
-        const errors = await validate(object);
-        if (errors.length > 0)
-            throw new BadRequestException("validation failed");
-        return value;
-    }
-
-    private toValidate(metatype : Function) : boolean {
-        const types : Function[] = [String, Boolean, Number, Array, Object ]
-        return !types.includes(metatype);
-    }
-}
