+++ backend/src/restRequests/auth/auth.module.ts
@@ -4,12 +4,11 @@ import { PassportModule } from '@nestjs/passport';
 import { AuthController } from './auth.controller';
 import { AuthService } from './auth.service';
 import { jwtConstants } from 'src/config/jwt.config';
-import { JwtStrategy } from './strategy/jwt.strategy';
-import { FortyTwoStrategy } from './strategy/fortytwo.strategy';
-import { RefreshStrategy } from './strategy/refresh.strategy';
+import { JwtStrategy } from 'src/strategy/jwt.strategy';
+import { FortyTwoStrategy } from 'src/strategy/fortytwo.strategy';
+import { RefreshStrategy } from 'src/strategy/refresh.strategy';
 import { Logger } from '@nestjs/common';
-import { JwtAuthenticationStrategy } from './strategy/jwt-2fa.strategy';
-import { UnauthorizedExceptionFilter } from 'src/filter/UnauthorizedExceptionFilter';
+import { JwtAuthenticationStrategy } from 'src/strategy/jwt-2fa.strategy';
 
 @Module({
   imports : [
@@ -19,7 +18,7 @@ import { UnauthorizedExceptionFilter } from 'src/filter/UnauthorizedExceptionFil
       signOptions : {expiresIn : jwtConstants.access_expiresIn}}),
   ],
   controllers: [AuthController],
-  providers: [UnauthorizedExceptionFilter, Logger, AuthService, JwtStrategy, JwtAuthenticationStrategy, FortyTwoStrategy, RefreshStrategy],
+  providers: [Logger, AuthService, JwtStrategy, JwtAuthenticationStrategy, FortyTwoStrategy, RefreshStrategy],
   exports : [JwtModule, Logger, AuthService]
 })
